Significant update to enable HIP cmake build
The main key update has been to the cmake files. Now this code will run cmake happily with newer versions of rocm. Current tests with <=5.0.2 all failed but 5.4.3 passed. This has to do with issues in older rocms but exaclty what I have not dived into. An example of the issues seen with older rocm is the same as those reported in https://gitlab.kitware.com/cmake/cmake/-/issues/23339

- includes minor update of logging, memory reporting, parallel API   report based on the github.com/pelahi/profile_util repo
- includes a silly work around as cmake build system when using hip will ignore files with the .cu and assume such files have a .hip suffix. To overcome this issue I have just generated soft links to the cu files. Eventually the idea would be to remove the permenant soft links in the repo and just add a macro to generate them upon configuration.